protobuf without protoc
