+++ recipes/protobuf/3.9.x/CMakeLists.txt
@@ -1,6 +1,7 @@
 cmake_minimum_required(VERSION 2.8.12)
 project(cmake_wrapper)
 
+message(STATUS "Conan CMake Wrapper")
 include("conanbuildinfo.cmake")
 conan_basic_setup(TARGETS)
